LPS-113473 Move EmptyState to correct folder
