+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-report-web/src/main/resources/META-INF/resources/js/components/card/Card.es.js
@@ -17,7 +17,7 @@ import ClayIcon from '@clayui/icon';
 import {ClayTooltipProvider} from '@clayui/tooltip';
 import React from 'react';
 
-import EmptyState from './EmptyState.es';
+import EmptyState from '../empty-state/EmptyState.es';
 
 export default ({children, fieldName, totalEntries, type}) => (
 	<div className="col-md-8">
