+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-report-web/src/main/resources/META-INF/resources/js/index.es.js
@@ -14,8 +14,8 @@
 
 import React from 'react';
 
-import EmptyState from '../js/components/card/EmptyState.es';
 import App from './App.es';
+import EmptyState from './components/empty-state/EmptyState.es';
 
 export default ({data, fields}) => {
 	if (!data || data.length === 0) {
